Fixes compilation issues in tests
